@@ -99,25 +99,29 @@ dependencies:
 #
   #  geocoding: ^2.0.2
   geocoding: ^2.1.0
-  google_maps_flutter: ^2.2.5
-  social_share: ^2.1.1
-  firebase_dynamic_links: ^5.0.16
-  wc_flutter_share: ^0.4.0
-  permission_handler: ^10.2.0
-  onesignal_flutter: ^3.0.0
-  package_info: ^2.0.2
-  device_info: ^2.0.2
-  flutter_background: ^1.0.2+1
-  flutter_background_service: ^2.4.4
-  flutter_background_service_android: ^3.0.2
-  flutter_local_notifications: ^12.0.2
-  flutter_cache_manager: ^3.1.2
-  connectivity: ^3.0.6
-  flutter_gifimage: ^1.0.0
+  google_maps_flutter: ^2.3.1
+  social_share: ^2.3.1
+  firebase_dynamic_links: ^5.3.3
+#  wc_flutter_share: ^0.4.0 (discontinued replaced by share_plus)
+  permission_handler: ^10.3.0
+  onesignal_flutter: ^3.5.1
+#  package_info: ^2.0.2 discontinued replaced by package_info_plus
+#  device_info: ^2.0.3 discontinued replaced by device_info_plus
+  device_info_plus: ^9.0.2
+  flutter_background: ^1.2.0
+  flutter_background_service: ^3.0.1
+  flutter_background_service_android: ^4.0.1
+  flutter_local_notifications: ^14.1.1
+#  flutter_cache_manager: ^3.3.0 problem with http
+#  connectivity: ^3.0.6 discontinued replaced by connectivity_plus
+#  flutter_gifimage: ^1.0.0 problem with null safety (safely can be removed)
   custom_info_window: ^1.0.1
-  share: ^2.0.4
-  gallery_saver: ^2.3.2
-  webview_flutter: ^4.0.1
+#  share: ^2.0.4 discontinued replaced by share_plus
+  share_plus: ^7.0.2
+#  gallery_saver: ^2.3.2 problem problem with http & replaced by image_galler_saver
+  image_gallery_saver: ^2.0.2 # alternative of gallery_saver
+  dio: ^5.2.1+1
+  webview_flutter: ^4.2.2
 #  carp_background_location: ^3.0.1
 
   ## End of packages supported by only "Android"##
