@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import '/constants/colors.dart';
 import '/constants/spaces.dart';
-import 'package:flutter_polyline_points/flutter_polyline_points.dart';
+// import 'package:flutter_polyline_points/flutter_polyline_points.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import '/functions/mapUtils/getLoactionUsingImei.dart';
 import '/widgets/Header.dart';
@@ -22,7 +22,7 @@ class HistoryScreen extends StatefulWidget {
 }
 
 class _HistoryScreenState extends State<HistoryScreen> {
-  late final List<PointLatLng> polylinePoints;
+  late final List<LatLng> polylinePoints;
   late GoogleMapController _googleMapController;
   late LatLng lastlatLngMarker;
   Iterable markers = [];
